AI_Research_assistant_project/requirements.txt
openai==1.3.9
pinecone_client==2.0.13
tiktoken
tqdm==4.64.1
kaggle==1.5.13
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.29.0
faiss-cpu==1.7.4
streamlit-extras
altair<5
googletrans==3.1.0a0
langchain
langdetect==1.0.9
